renamed prototype to polyfill
